+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Allow configuring loadBalancerIP for LoadBalancer type services
 * Allow setting labels and annotations for Persistent Volume Claims
 * Add support for Jaeger tracing in Kafka Mirror Maker and Kafka Connect
+* Add support for deploying Kafka Exporter
 
 ## 0.13.0